refactor(visual-tests): move single-owner demos next to their components
A demo beside its component moves, gets reviewed and rots together with it.
Applied only where exactly one component owns the demo: Chip, AppNotifications
and the hangar SyncBtn result panel.

The rest stay central for a reason. A page covering a whole family has no
single owner - buttons.vue documents Btn, BtnGroup and BtnDropdown, so filing
it under Btn/ would hide the other two from whoever changes them. Same for
tables.vue and panels.vue. The composed pages have no owner at all.

support-hint.vue renders the hint in its real context, inside the sync result
panel, so it imports StatePreview from a frontend/ component. Co-locating it
under shared/components/SupportHint/ would make shared/ depend on frontend/.

Route names and paths are unchanged, so only the route index moved with them.
