+++ app/frontend/frontend/components/Hangar/SyncBtn/Result/visual.vue
@@ -149,7 +149,7 @@ const comlink = useComlink();
 const openState = (state: State) => {
   comlink.emit("open-modal", {
     component: () =>
-      import("@/frontend/pages/visual-tests/sync-modal/StatePreview.vue"),
+      import("@/frontend/components/Hangar/SyncBtn/Result/visual/StatePreview.vue"),
     props: {
       title: `Hangar Sync — ${state.label}`,
       processSteps: state.processSteps,
