+++ app/frontend/frontend/components/Hangar/SyncBtn/Result/visual/StatePreview.vue
