+++ app/frontend/shared/components/AppNotifications/visual/ImageDemo.vue
