+++ app/frontend/shared/components/base/Chip/visual.vue
